Fix underflow bug with msgpack_size_limit chunking logic
Found when running analysis with Curosr on the msgpack data and finding a solution. The initial thoght was we had an issue with nested dynamic shapes, but this isn't the case here. The underflow makes it appear we had more data than nessicary to pack causing EOF and other sorts of packing issues